Replaced custom exception with native RuntimeError
Rmoved custom RuntimeException class that extends RuntimeError.

When multiple process are in the same application group and the
:multiple option is not set, the application should fail with a run time
error. This was not happening with the custom exception and was only
printing out a warrning.

If :multiple is false and multiple applications are in the same app
group - not necessarily running - then I want the application to fail
and exit throwing an error.